requirements.txt
opencv-python
hidapi
pytest
pytest-mock
